requirements.txt
awscli==1.16.42
beautifulsoup4==4.6.3
boto3==1.9.32
botocore==1.12.32
bottlenose==1.1.8
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==1.7.1
docutils==0.14
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
keyring==10.1
keyrings.alt==1.3
lxml==4.2.5
MarkupSafe==1.0
pyasn1==0.4.4
pycrypto==2.6.1
pygobject==3.22.0
python-apt==1.1.0b5
python-dateutil==2.7.4
python-dotenv==0.9.1
pyxdg==0.25
PyYAML==3.13
requests==2.20.0
rsa==3.4.2
s3transfer==0.1.13
SecretStorage==2.3.1
six==1.11.0
ssh-import-id==5.6
unattended-upgrades==0.1
urllib3==1.24
Werkzeug==0.14.1
